+++ CHANGELOG.md
@@ -21,3 +21,4 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 - Removed the feature that made the Enderman's staring sound move along with it.
   - This was simply out of the mod's scope and detracted from the Enderman's fear factor.
 - Added `fingerspit`'s new music tracks to Wilder Wild's custom music pools, thanks to `Rebel459`!
+- Removed the `wilderwild:beta_beach_enabled` and `wilderwild:snow_under_mountain_enabled` `ConditionSources`, as they have been made redundant by FrozenLib's `ConfigPredicate`s.